ai-service/models/recommendation_model.py
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.decomposition import TruncatedSVD
import pandas as pd
import numpy as np
import joblib
import os

class RecommendationModel:
    """
    Hybrid recommendation engine combining Collaborative Filtering (SVD) 
    and Content-Based Filtering (Cosine Similarity).
    """
    def __init__(self):
        self.model_path = 'assets/recommender.pkl'
        self.is_trained = os.path.exists(self.model_path)
        if self.is_trained:
            self.model_data = joblib.load(self.model_path)
        else:
            self.model_data = None

    def get_personalized(self, user_id, user_history, product_catalog):
        """
        Generate top N personalized recommendations.
        Fallbacks to content-based or popularity if data is sparse.
        """
        # Logic for research-grade hybrid filtering
        # 1. Collaborative: Find similar users using SVD decomposition of interaction matrix
        # 2. Content: Find similar products using NLP features of descriptions
        # 3. Hybrid: Weighted average of both scores
        
        # MOCK IMPLEMENTATION (Simulates AI behavior for group-based customization)
        # Returns a subset of products sorted by 'relevance'
        
        # In a real experiment, weights would change based on groupType
        sorted_products = sorted(product_catalog, key=lambda x: x.get('rating', 0), reverse=True)
        return sorted_products[:10]

    def get_similar(self, product_id, product_catalog, n=5):
        """
        Finds products similar to a target item using content-based similarity.
        """
        return product_catalog[:n] # Placeholder
