ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.15.0
scikit-learn==1.3.2
textblob==0.17.1
vaderSentiment==3.3.2
numpy==1.26.2
pandas==2.1.3
python-multipart==0.0.6
pydantic==2.5.2
joblib==1.3.2
face-recognition==1.3.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
